Efficient Q-Q plot

Density-like QQ-plot with hexagons for high-dimensional data.

stat_gwas_qq_hex(mapping = NULL, data = NULL, geom = "hex",
  position = "identity", na.rm = FALSE, bins = 30, binwidth = NULL,
  show.legend = NA, inherit.aes = TRUE, y.thresh = NULL,
  hex.function = hexBinSummarise, fill = "black", ...)

geom_gwas_qq_hex(mapping = NULL, data = NULL, geom = "hex",
  position = "identity", na.rm = FALSE, bins = 30, binwidth = NULL,
  show.legend = NA, inherit.aes = TRUE, y.thresh = NULL,
  hex.function = hexBinSummarise, fill = "black", ...)

Arguments

mapping
Set of aesthetic mappings created by aes() or aes_(). If specified and inherit.aes = TRUE (the default), it is combined with the default mapping at the top level of the plot. You must supply mapping if there is no plot mapping.

data
The data to be displayed in this layer. There are three options:

If NULL, the default, the data is inherited from the plot data as specified in the call to ggplot().

A data.frame, or other object, will override the plot data. All objects will be fortified to produce a data frame. See fortify() for which variables will be created.

A function will be called with a single argument, the plot data. The return value must be a data.frame, and will be used as the layer data. A function can be created from a formula (e.g. ~ head(.x, 10)).

geom
Override the default connection between geom_hex and stat_binhex.

position
Position adjustment, either as a string, or the result of a call to a position adjustment function.

na.rm
If FALSE, the default, missing values are removed with a warning. If TRUE, missing values are silently removed.

bins
numeric vector giving number of bins in both vertical and horizontal directions. Set to 30 by default.

binwidth
Numeric vector giving bin width in both vertical and horizontal directions. Overrides bins if both set.

show.legend
logical. Should this layer be included in the legends? NA, the default, includes if any aesthetics are mapped. FALSE never includes, and TRUE always includes. It can also be a named logical vector to finely select the aesthetics to display.

inherit.aes
If FALSE, overrides the default aesthetics, rather than combining with them. This is most useful for helper functions that define both data and aesthetics and shouldn't inherit behaviour from the default plot specification, e.g. borders().

y.thresh
Same scale as y (e.g. 0.05), y <= y.thresh AFTER computing expected.

hex.function
ggGWAS:::hexBinSummarise or ggplot2:::hexBinSummarise

fill
color by which hexagons are filled, by default black.

...
Other arguments passed on to layer(). These are often aesthetics, used to set an aesthetic to a fixed value, like colour = "red" or size = 3. They may also be parameters to the paired geom/stat.

Details

Code and documentation mostly from https://github.com/tidyverse/ggplot2/blob/master/R/stat-binhex.r.

See also

Examples

require(ggplot2)

n.sample <- 1000
df <- data.frame(P = runif(n.sample), GWAS = sample(c("a", "b"),
  n.sample,
  replace = TRUE
))
theme_set(theme_bw())

(qp <- ggplot(df, aes(y = P)) +
  stat_gwas_qq_hex() +
  geom_abline(intercept = 0, slope = 1))

(qp <- ggplot(df, aes(y = P, group = GWAS, color = GWAS)) +
  stat_gwas_qq_hex() +
  geom_abline(intercept = 0, slope = 1))